Adjustment with merge
