--- client/app/components/dashboard/ualReport/ualDataSource/ualDataSource.html
@@ -1,20 +1,0 @@
-<ual-data-source>
-    <div class="top">
-        <i class="ion-ios-close-empty" ng-click="vm.close()"></i>
-    </div>
-
-    <div class="container">
-        <div class="title">
-            Select Data Source
-        </div>
-        <div class="content">
-            <input type="text" ng-model="vm.searchFilter">
-            <ual-data-source-list search-filter="vm.searchFilter"></ual-data-source-list>
-
-        </div>
-
-        <div class="buttons">
-            <ual-button class="-primary">Apply</ual-button>
-        </div>
-    </div>
-</ual-data-source>
