--- client/app/components/dashboard/ualReport/ualDataSource/ualDataSourceItem/ualDataSourceItem.controller.js
@@ -1,12 +1,0 @@
-class UalDataSourceItemController {
-  /*@ngInject*/
-  constructor() {
-      this.filteredElements = [];
-  }
-  hasElements(){
-      return !!this.filteredElements &&  this.filteredElements.length > 0;
-  }
-  
-}
-
-export default UalDataSourceItemController;
