+++ client/app/components/dashboard/ualReport/ualReportForm/ualDataSource/ualDataSourceItem/ualDataSourceItem.scss
@@ -1,4 +1,4 @@
-@import '../../../../../common/common';
+@import '../../../../../../common/common';
 
 ual-data-source-item {
     .data-source-name{
