+++ client/app/components/dashboard/ualReport/ualReportForm/ualDataSource/ualDataSourceList/ualDataSourceList.scss
@@ -1,4 +1,4 @@
-@import '../../../../../common/common';
+@import '../../../../../../common/common';
 
 //Layout
 ual-data-source-list {
